fastapi-job-recommender/pyproject.toml
[tool.poetry]
name = "fastapi-job-recommender"
version = "0.1.0"
description = "A FastAPI application for job recommendations based on user prompts."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.0"
uvicorn = "^0.17.0"
pandas = "^1.3.0"
openai = "^0.11.0"
faiss-cpu = "^1.7.0"
python-dotenv = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"